+++ projects/onlyoffice/document-editor-angular/src/lib/components/document-editor.component.ts
@@ -1,5 +1,5 @@
 /*
-* (c) Copyright Ascensio System SIA 2024
+* (c) Copyright Ascensio System SIA 2025
 *
 * Licensed under the Apache License, Version 2.0 (the "License");
 * you may not use this file except in compliance with the License.
@@ -30,11 +30,13 @@ declare global {
   selector: 'document-editor',
   template: '<div [id]="id"></div>',
   styles: [
-  ]
+  ],
+  standalone: false
 })
 export class DocumentEditorComponent implements OnInit, OnChanges, OnDestroy {
   @Input() id: string;
   @Input() documentServerUrl: string;
+  @Input() shardkey: string | boolean = true;
   @Input() config: IConfig;
 
   @Input() document_fileType?: string;
@@ -84,8 +86,16 @@ export class DocumentEditorComponent implements OnInit, OnChanges, OnDestroy {
     let url = this.documentServerUrl;
     if (!url.endsWith("/")) url += "/";
 
-    const docApiUrl = `${url}web-apps/apps/api/documents/api.js`;
-    loadScript(docApiUrl, "onlyoffice-api-script")
+    let docsApiUrl = `${url}web-apps/apps/api/documents/api.js`;
+    if (this.shardkey) {
+      if (typeof this.shardkey === "boolean") {
+        docsApiUrl += `?shardkey=${this.config.document?.key}`;
+      } else {
+        docsApiUrl += `?shardkey=${this.shardkey}`;
+      }
+    }
+
+    loadScript(docsApiUrl, "onlyoffice-api-script")
       .then(() => this.onLoad())
       .catch((err) => {
         this.onError(-2);