+++ src/app/app.component.ts
@@ -1,10 +1,12 @@
 import { Component } from '@angular/core';
-import { IConfig } from '@onlyoffice/document-editor-angular';
+import { DocumentEditorModule, IConfig } from '@onlyoffice/document-editor-angular';
 import config from "./../../config/config.json";
+import { CommentsComponent } from './components/comments/comments.component';
 
 @Component({
   selector: 'app-root',
-  templateUrl: './app.component.html'
+  templateUrl: './app.component.html',
+  imports: [ DocumentEditorModule, CommentsComponent ]
 })
 export class AppComponent {
   comments: any[] = [];